% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{clean_string}
\alias{clean_string}
\title{Clean and standardize strings}
\usage{
clean_string(x)
}
\arguments{
\item{x}{Character vector to clean}
}
\value{
Cleaned character vector
}
\description{
Clean and standardize strings
}
\keyword{internal}
