% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{import_data}
\alias{import_data}
\title{Import Insurance Data from Excel Files}
\usage{
import_data(customers_path, policies_path, claims_path)
}
\arguments{
\item{customers_path}{Character string Path to Excel file containing customer data}

\item{policies_path}{Character string. Path to Excel file containing policy data}

\item{claims_path}{Character string. Path to Excel file containing claims data}
}
\value{
List containing three data frames:
  \describe{
    \item{customers}{Data frame of customer information}
    \item{policies}{Data frame of policy information}
    \item{claims}{Data frame of claims information}
  }
}
\description{
Reads customer, policy, and claims data from excel files and returns them as list of data frames
}
\examples{
\dontrun{
data <- import_data(
  customers_path = "data/customers.xlsx",
  policies_path = "data/policies.xlsx",
  claims_path = "data/claims.xlsx"
)
}
}
