% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{impute_median}
\alias{impute_median}
\title{Impute median values with bounds checking}
\usage{
impute_median(x, min_val = -Inf, max_val = Inf)
}
\arguments{
\item{x}{Numeric vector to impute}

\item{min_val}{Minimum allowed value}

\item{max_val}{Maximum allowed value}
}
\value{
Imputed numeric vector
}
\description{
Impute median values with bounds checking
}
\keyword{internal}
