% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{replace_unknown}
\alias{replace_unknown}
\title{Replace unknown values with standardized "unknown"}
\usage{
replace_unknown(x)
}
\arguments{
\item{x}{Character vector to process}
}
\value{
Character vector with standardized unknown values
}
\description{
Replace unknown values with standardized "unknown"
}
\keyword{internal}
