% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\docType{data}
\name{UNKNOWN_VALUES}
\alias{UNKNOWN_VALUES}
\title{Constants used across prprocessing functions}
\format{
An object of class \code{character} of length 5.
}
\usage{
UNKNOWN_VALUES
}
\description{
Constants used across prprocessing functions
}
\keyword{internal}
