% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{validate_processed_data}
\alias{validate_processed_data}
\title{Validate processed insurance data}
\usage{
validate_processed_data(processed_data)
}
\arguments{
\item{processed_data}{List of processed data frames}
}
\value{
TRUE if validation passes, errors if not
}
\description{
Validate processed insurance data
}
\keyword{internal}
